pyproject.toml
[tool.poetry]
name = "sistema-perdao-faltas"
version = "0.1.0"
description = ""
authors = ["guguRamos <batista.gustavo@aluno.ifsp.edu.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
asgiref = "^3.8.1"
django-cors-headers = "^4.6.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
pyjwt = "^2.10.1"
pytz = "^2024.2"
sqlparse = "^0.5.3"
mysqlclient = "^2.2.7"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
